requirements.txt
discord.py
python-dotenv
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
